requirements.txt
Pillow>=10.0.0
imageio>=2.33.1
pytest>=8.0.0
numpy>=1.26.4 